Small tweaks here and there. Changed unsigned to size_t where it's technically correct. Improved wording for some comments. A couple of small simplifications and better naming. A default parameter where there should have been.

+++ src/GNDStk/Node/src/one.hpp
@@ -37,7 +37,7 @@ const Node &one(
    // name: #times node name == key
    // both: #times node name == key AND the filter condition is met
    const Node *theone = nullptr;
-   unsigned name = 0, both = 0;
+   size_t name = 0, both = 0;
    for (auto &c : children)
       if (c->name == key && (name++, filter(*c)) && both++ == 0)
          theone = &(*c);
